\name{biGraph-internal}
\alias{covLoss}
\alias{covLoss.single.edge}
\alias{vertex.avCovLoss}
\alias{collapse.bigzList}
\alias{graph.dH}
\alias{vertex.dH}
\alias{edge.dH}
\alias{get.Hafter.vertex}
\alias{get.Hafter.edge}
\alias{redundant.cliques}
\alias{redundant.cliques.parallel}
\alias{number.subcliques}
\alias{jaccard.distance.single.vertex}
%%\docType{package}
\title{
Internal functions of the biGraph--package
}
\description{
These functions are intended for internal use only and are not supposed to be called by a user directly.
}
\author{
Ingo Vogt   \email{bigraph.vogt@gmail.com}
}
\references{
Vogt, I. and Mestres, J. (in preparation) Assessing information loss associated with bipartite network projection.
}
\keyword{ biGraph }
\seealso{
\code{\link[igraph]{igraph}}
}
\examples{
}